Remove CSS var defaults from theme-ui to fix SSR mismatch
This is a partial application of https://github.com/system-ui/theme-ui/pull/1453/files

Note that if I ever want to use @theme-ui/color, I'll need to apply more
of that PR.

The problem was that when mode == "auto" and `prefers-color-scheme: dark`,
you get an SSR mismatch.  This is because some components use the color
values in their CSS.  On the server this value was `var(--blah, #123)`,
but on the client the value is `var(--blah, #456)` because theme-ui
includes the real value as a default argument.  This is for use in
@theme-ui/color but is otherwise superfluous.  That difference causes a
difference in classnames.

Since our CSS variables change *for* us, we can strip the default arg so
the CSS hashes the same way regardless of mode.  There might be some other
kind of internal inconsistencies happening in this situation, but it fixes
the error so I'm happy.