Fix GC bug, move GC pixel to _document, split file
The pixel was not actually showing up in prod because the VERCEL_URL can
never equal "ahurle.dev".  I replaced this with a check for
VERCEL_GIT_REPO_OWNER.

I also realized that there's no need for the client to hydrate the
<noscript> tag and keep it up to date, so I moved it to _document in order
to only render on the server. I split the file apart to avoid getting
confused about server/client environment